several updated
Built in several safety mechanisms: stop function for all axes which is called when an exception is raised. Exception is also raised when error on motion controller occurs. Motion can now be called with seperate perform_motion() function. Monitoring of motions is cleaned up. percentage done is now calculated correctly.